requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify-client>=1.4.0
requests>=2.31.0
aiohttp>=3.9.0
python-dateutil>=2.8.2
validators>=0.20.0
crewai>=0.11.0
python-dotenv>=0.19.0
nest-asyncio>=1.5.6
